recent trade component added
The scripts can now get OB and recent trades, then output a few STATS:
Task now is to connect the two meaning full data together. A new functions will be required.
The end goal here is to create a powerful script that can service as an algorithmic trading tool for better trade diligence.